# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "KnxMonitor.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Recording/settings": {
      "get": {
        "tags": [
          "Recording"
        ],
        "summary": "Current recording settings: hot-buffer size and cold-archive options.",
        "description": "Served from the cached snapshot that the recording services themselves read, so this needs no database round-trip.",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Recording"
        ],
        "summary": "Updates the recording settings; they take effect on the running recording.",
        "description": "The values are persisted and swapped into the snapshot the recording workers read, so no restart is needed. What the fields do: HotBufferMaxCount is the number of telegrams kept in the database — a retention pass periodically deletes everything older than the newest HotBufferMaxCount rows, so lowering the value discards the oldest telegrams. Deletion happens chunked on the following passes, not the instant the value is saved. ArchiveEnabled toggles the additional NDJSON+gzip day-files; switching it off stops new writes but leaves existing archive files untouched. ArchiveRetentionDays deletes archive day-files older than that many days, null keeps them forever. The sweep runs at startup and at the daily file rollover, so a shortened retention only takes effect at the next sweep. Validation: HotBufferMaxCount must be greater than 0, ArchiveRetentionDays must be null or greater than 0 — otherwise 400 and nothing is changed.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordingSettingsDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordingSettingsDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RecordingSettingsDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The settings as they are now in effect."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RecordingSettingsDto": {
        "type": "object",
        "properties": {
          "hotBufferMaxCount": {
            "type": "integer",
            "format": "int32"
          },
          "archiveEnabled": {
            "type": "boolean"
          },
          "archiveRetentionDays": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Access-Token aus POST /api/Auth/login (ohne \"Bearer \"-Präfix einfügen).",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}